tratando erro 404

+++ src/main/java/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -58,24 +59,39 @@ public class TopicosController {
 	}
 
 	@GetMapping("/{id}") // pega o Id pela url
-	public DetalhesTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getById(id);
-		DetalhesTopicoDto dto = new DetalhesTopicoDto(topico);
-		return dto;
+	public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {
+			 return ResponseEntity.ok(new DetalhesTopicoDto(topico.get())); 
+		}
+		
+		return ResponseEntity.notFound().build();
+			
 	}
 
 	@PutMapping("/{id}")
 	// dispara o update no banco de dados
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-		Topico topico = form.atualizar(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+			  
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
-
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+			  
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 }
